Refactor page structure
